test: stabilize current live CI failures
